refactor(canvas): Keyboard-Shortcuts aus CanvasArea in Hook auslagern (#468)
CanvasArea.tsx 2016 → 1898 Zeilen. Der ~120-zeilige window-keydown-Handler
(Esc, Ctrl/Cmd+C/V/D, +/=, Pfeil-Nudge, Delete/Backspace) → neuer Hook
useCanvasKeyboardShortcuts. Eine abgegrenzte Concern (Event-Lifecycle +
Key→Aktion-Mapping); die Fähigkeiten werden injiziert (Store-Action-Typen
via ProjectState[...], damit das Interface exakt passt).

Verhaltensneutral, golden-bewiesen (Characterization-Test mit echter
Tastatur-Interaktion):
- Nudge: ArrowRight×2 → +22px (2×Grid), identisch vorher/nachher.
- Duplicate: Ctrl+D → Node 2→3, identisch.
- Delete: → 3→2, identisch. 0 Console-Errors.
- tsc 0, eslint 0, build 0.

Refs #468
